Addsingleton. AddScoped, AddTransient
AddSingleton - Instance is created once the service is first requested and the same instance is used for all http requests through out the execution.

AddScoped: Every http request gets a new instance. We get the same instance within the scope of a given http request but a new instance across different http requests. Everyt ime the constructor is called for instance.

AddTransient: A new instance is provided every time. Irrespective of scope.